Created a Mixin class for permissions
We did this for dry property and also avoid having permissions changed in n diff locations if this could be done in a single mixin class
