+++ backend/api/mixin.py
@@ -1,0 +1,5 @@
+from .permissions import IsStaffEditorPermission
+from rest_framework import permissions
+
+class StaffEditorPermissionMixin:
+    permission_classes = [permissions.IsAdminUser, IsStaffEditorPermission]
